- parse fasta file as input - parse JSON as input for rates and characters - straight line evolution of nucleotides and character
Let's say the time is 1 My, we have a sequence of ten letters and one character. All three rates are given in units of changes per nucleotide per My. This puts all the rates on the scale, but note that the character rate is affected by the locus length. To go to the unaffected rate, need to divide by the length of the locus.

$\lambda = 1$ means 1 character change and 0 letter changes expected per nucleotide (10 letter changes)
$\mu = 1$ means 1 letter change expected per nucleotide (10 letter changes in total)
$\nu = 1$ means 1 letter **and** 1 character change expected per nucleotide (10 letter changes and 10 character changes)

High $\lambda$ and high $\mu$ together is undistinguishable from high $\nu$. For this reason we need to  always start from some value of $\nu$ and then $\lambda$ or $\mu$ on top.

For example instead of

$\lambda=0.2$            $\lambda = 0.1$
$\mu=0.1$    =>   $\mu=0$
$\nu=0.1$            $\nu = 0.2$